// prodis-prodiswcl/src/app/modules/prodisapi/model/datiAzienda.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Atecofin } from './atecofin';
import { Ccnl } from './ccnl';
import { Comune } from './comune';
import { Dichiarante } from './dichiarante';
import { Prospetto } from './prospetto';
import { SedeLegale } from './sedeLegale';
import { StatiEsteri } from './statiEsteri';


export interface DatiAzienda { 
    atecofin?: Atecofin;
    capReferente?: string;
    ccnl?: Ccnl;
    cfAzienda?: string;
    cfCapogruppo?: string;
    cfReferente?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    cognomeReferente?: string;
    comune?: Comune;
    dAggiorn?: Date;
    dInserim?: Date;
    denominazioneDatoreLavoro?: string;
    dichiarante?: Dichiarante;
    elencoSedi?: Array<SedeLegale>;
    emailReferente?: string;
    faxReferente?: string;
    flgCapogruppoEstero?: string;
    flgProspettoDaCapogruppo?: string;
    id?: number;
    idAziendaSilp?: string;
    indirizzoReferente?: string;
    nomeReferente?: string;
    prospetto?: Prospetto;
    sedeLegale?: SedeLegale;
    statiEsteri?: StatiEsteri;
    telefonoReferente?: string;
}