// prodis-prodiswcl/src/app/modules/prodisapi/model/datiProvinciali.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CategorieEscluse } from './categorieEscluse';
import { LavoratoriInForza } from './lavoratoriInForza';
import { PartTime } from './partTime';
import { PostiLavoroDisp } from './postiLavoroDisp';
import { ProspettoProvSede } from './prospettoProvSede';
import { ProspettoProvincia } from './prospettoProvincia';
import { ProvCompensazioni } from './provCompensazioni';
import { ProvConvenzione } from './provConvenzione';
import { ProvEsonero } from './provEsonero';
import { ProvEsoneroAutocert } from './provEsoneroAutocert';
import { ProvGradualita } from './provGradualita';
import { ProvIntermittenti } from './provIntermittenti';
import { ProvSospensione } from './provSospensione';


export interface DatiProvinciali { 
    baseComputo?: number;
    categorieEscluses?: Array<CategorieEscluse>;
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    flgAziConConvenzione?: string;
    flgAziEsoneroParziale?: string;
    flgAziGradu?: string;
    flgAziSosp?: string;
    flgCategorieEscluseCompute?: string;
    flgCompensazioniTerritorial?: string;
    flgDettaglioIntermittenti?: string;
    flgDettaglioPartime?: string;
    flgLavoratoriInForza?: string;
    flgPostiLavoroDisponibili?: string;
    id?: number;
    lavoratoriInForzas?: Array<LavoratoriInForza>;
    nCateProtForza?: number;
    nCateProtForzaA17012000?: number;
    nCateProtForzaCntDis?: string;
    nCateProtForzaEsubero?: string;
    nCentralTelefoNonvedenti?: number;
    nConvenzioni12bis14Ft?: number;
    nDisabiliInForza?: number;
    nIntermittentiRiproporziona?: number;
    nPartimeRiproporzionati?: number;
    nPostiPrevCentraliNonved?: number;
    nPostiPrevMassofisNonved?: number;
    nSomministratiFt?: number;
    nTelelavoroFt?: number;
    nTerariabMassofisNonved?: number;
    nTotaleLavoratDipendenti?: number;
    note?: string;
    partTimes?: Array<PartTime>;
    postiLavoroDisps?: Array<PostiLavoroDisp>;
    prospettoProvSede?: ProspettoProvSede;
    prospettoProvincia?: ProspettoProvincia;
    provCompensazionis?: Array<ProvCompensazioni>;
    provConvenzione?: ProvConvenzione;
    provEsonero?: ProvEsonero;
    provEsoneroAutocert?: ProvEsoneroAutocert;
    provGradualita?: ProvGradualita;
    provIntermittentis?: Array<ProvIntermittenti>;
    provSospensione?: ProvSospensione;
}