// prodis-prodiswcl/src/app/modules/prodisapi/model/prospetto.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssPubbliche } from './assPubbliche';
import { CategoriaAzienda } from './categoriaAzienda';
import { Comunicazione } from './comunicazione';
import { DatiAzienda } from './datiAzienda';
import { ProspettoGradualita } from './prospettoGradualita';
import { ProspettoProvincia } from './prospettoProvincia';
import { RiepilogoNazionale } from './riepilogoNazionale';
import { Soggetti } from './soggetti';
import { StatoProspetto } from './statoProspetto';
import { StatoVerifica } from './statoVerifica';


export interface Prospetto { 
    annoProtocollo?: number;
    assPubbliche?: Array<AssPubbliche>;
    categoriaAzienda?: CategoriaAzienda;
    cfComunicazione?: string;
    cfStudioProfessionale?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    codiceComunicazione?: string;
    codiceComunicazionePreced?: string;
    comunicazione?: Comunicazione;
    dAggiorn?: Date;
    dFineSospensioneQ1?: Date;
    dInserim?: Date;
    dataInvio?: Date;
    dataPrimaAssunzione?: Date;
    dataProtocollo?: Date;
    dataRiferimentoProspetto?: Date;
    dataSecondaAssunzione?: Date;
    dataTimbroPostale?: Date;
    datiAzienda?: DatiAzienda;
    emailNotifica?: string;
    emailSoggettoComunicazione?: string;
    flgAssunzioniPubbSelezione?: string;
    flgConfermatoQ1?: string;
    flgGradualita?: string;
    flgInvioMinistero?: string;
    flgNessunaAssunzioneAggiun?: string;
    flgSospensionePerMobilita?: string;
    flgVisitaIspettiva?: string;
    id?: number;
    note?: string;
    numLavorInForzaNazionale?: number;
    numeroProtocollo?: number;
    idProspettoPrecedente?: number;
    prospettoGradualita?: ProspettoGradualita;
    prospettoProvincias?: Array<ProspettoProvincia>;
    riepilogoNazionale?: RiepilogoNazionale;
    soggetti?: Soggetti;
    statoProspetto?: StatoProspetto;
    statoVerifica?: StatoVerifica;
    tipoProvenienza?: string;
}